Add abstract UNIX socket handling

+++ src/socket/unix_server.cr
@@ -24,17 +24,18 @@ class UNIXServer < UNIXSocket
   # Creates a named UNIX socket, listening on a filesystem pathname.
   #
   # Always deletes any existing filesystam pathname first, in order to cleanup
-  # any leftover socket file.
+  #
+  # Note: An abstract UNIX server act on virtual files, thus not creating nor deleting anything.
   #
   # The server is of stream type by default, but this can be changed for
   # another type. For example datagram messages:
   # ```
   # UNIXServer.new("/tmp/dgram.sock", Socket::Type::DGRAM)
   # ```
-  def initialize(@path : String, type : Type = Type::STREAM, backlog = 128)
-    super(Family::UNIX, type)
+  def initialize(@path : String, type : Type = Type::STREAM, backlog = 128, abstract is_abstract = false)
+    super(Family::UNIX, type, is_abstract)
 
-    bind(UNIXAddress.new(path)) do |error|
+    bind(UNIXAddress.new(path, is_abstract)) do |error|
       close(delete: false)
       raise error
     end
@@ -49,8 +50,8 @@ class UNIXServer < UNIXSocket
   # server socket when the block returns.
   #
   # Returns the value of the block.
-  def self.open(path, type : Type = Type::STREAM, backlog = 128)
-    server = new(path, type, backlog)
+  def self.open(path, type : Type = Type::STREAM, backlog = 128, abstract is_abstract = false)
+    server = new(path, type, backlog, is_abstract)
     begin
       yield server
     ensure
@@ -64,7 +65,7 @@ class UNIXServer < UNIXSocket
   # this method.
   def accept? : UNIXSocket?
     if client_fd = accept_impl
-      sock = UNIXSocket.new(client_fd, type)
+      sock = UNIXSocket.new(client_fd, type, @abstract)
       sock.sync = sync?
       sock
     end
@@ -74,9 +75,9 @@ class UNIXServer < UNIXSocket
   def close(delete = true)
     super()
   ensure
-    if delete && (path = @path)
+    if !abstract? && delete && (path = @path)
       File.delete(path) if File.exists?(path)
-      @path = nil
     end
+    @path = nil
   end
 end
